Time Series Charts - Moved PieChart and more into charts/
