+++ src/app/charts/pie-chart.ts
@@ -1,4 +1,4 @@
-import colors from './colors.json';
+import colors from '../colors.json';
 import {
   Brolog,
 } from 'brolog';
