+++ src/app/geometry-ops.spec.ts
@@ -7,7 +7,7 @@ import {
   polygon as turfPolygon, FeatureCollection
 } from '@turf/helpers';
 import {Feature, lineString, LineString, multiLineString, Polygon, polygon} from '@turf/helpers';
-import EowDataCharts from './eow-data-charts';
+import EowDataCharts from './charts/eow-data-charts';
 import * as chai from 'chai';
 import GeometryOps from './geometry-ops';
 import Brolog from 'brolog';
